+++ crates/miden-testing/src/mock_chain/chain.rs
@@ -64,7 +64,7 @@ use crate::{MockChainBuilder, TransactionContextBuilder};
 /// # use anyhow::Result;
 /// # use miden_protocol::{
 /// #    account::auth::AuthScheme,
-/// #    asset::{Asset, FungibleAsset},
+/// #    asset::{Asset, AssetCallbackFlag, FungibleAsset},
 /// #    note::NoteType,
 /// # };
 /// # use miden_testing::{Auth, MockChain};
@@ -119,7 +119,7 @@ use crate::{MockChainBuilder, TransactionContextBuilder};
 ///     mock_chain
 ///         .committed_account(receiver.id())?
 ///         .vault()
-///         .get_balance(fungible_asset.faucet_id())?,
+///         .get_balance(fungible_asset.vault_key())?,
 ///     fungible_asset.amount()
 /// );
 /// # Ok(())
@@ -1221,7 +1221,7 @@ impl From<Account> for TxContextInput {
 #[cfg(test)]
 mod tests {
     use miden_protocol::account::auth::AuthScheme;
-    use miden_protocol::account::{AccountBuilder, AccountStorageMode};
+    use miden_protocol::account::{AccountBuilder, AccountType};
     use miden_protocol::asset::{Asset, FungibleAsset};
     use miden_protocol::note::NoteType;
     use miden_protocol::testing::account_id::{
@@ -1248,7 +1248,7 @@ mod tests {
     async fn private_account_state_update() -> anyhow::Result<()> {
         let faucet_id = ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET.try_into()?;
         let account_builder = AccountBuilder::new([4; 32])
-            .storage_mode(AccountStorageMode::Private)
+            .account_type(AccountType::Private)
             .with_component(BasicWallet);
 
         let mut builder = MockChain::builder();